requirements.txt
streamlit
openai
easyocr
pytesseract
pillow
transformers 
torch